+++ src/app/login/services/login.service.ts
@@ -1,0 +1,19 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class LoginService {
+  isLoggedIn: boolean = false;
+  isAdmin: boolean = false;
+
+  logIn(): boolean {
+    this.isLoggedIn = true;
+    return this.isAdmin;
+  }
+
+  logOut(): void {
+    this.isLoggedIn = false;
+    this.isAdmin = false;
+  }
+}