+++ src/app/products/models/product.interface.ts
@@ -1,11 +1,13 @@
 import { OperationSystem } from './operation-system.enum';
 
 export interface IProduct {
+  id: string | null;
+  img: string;
   name: string;
   description: string;
   price: number;
   isAvailable: boolean;
-  operationSystem: OperationSystem;
+  operationSystem?: OperationSystem;
   availableCount: number;
   itemsInCart: number;
 }