+++ src/app/products/product-routing.module.ts
@@ -1,0 +1,21 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, type Routes } from '@angular/router';
+import { ProductListComponent } from './components/product-list/product-list.component';
+import { ProductViewComponent } from './components/product-view/product-view.component';
+
+const routes: Routes = [
+  {
+    path: 'product-list',
+    component: ProductListComponent,
+  },
+  {
+    path: 'product/:productID',
+    component: ProductViewComponent,
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class ProductRoutingModule { }