+++ src/app/products/product.module.ts
@@ -2,17 +2,22 @@ import { NgModule } from '@angular/core';
 import { ProductComponent } from './components/product/product.component';
 import { ProductListComponent } from './components/product-list/product-list.component';
 import { SharedModule } from '../shared/shared.module';
+import { ProductRoutingModule } from './product-routing.module';
+import { ProductViewComponent } from './components/product-view/product-view.component';
 
 @NgModule({
   declarations: [
     ProductComponent,
     ProductListComponent,
+    ProductViewComponent,
   ],
     imports: [
-        SharedModule
+        SharedModule,
+        ProductRoutingModule
     ],
   exports: [
     ProductListComponent,
+    ProductViewComponent
   ],
 })
 export class ProductModule { }