+++ src/app/shared/shared.module.ts
@@ -5,6 +5,12 @@ import { StylizeDirective } from './directives/stylize/stylize.directive';
 import { OrderByPipe } from './pipes/order-by/order-by.pipe';
 import { LoaderComponent } from './components/loader/loader.component';
 import { FormsModule } from '@angular/forms';
+import { PathNotFoundComponent } from './components/path-not-found/path-not-found.component';
+import { GoBackComponent } from './components/go-back/go-back.component';
+import { GoToCartComponent } from './components/go-to-cart/go-to-cart.component';
+import { CartService } from '../cart/services/cart-service/cart.service';
+import { RouterLink } from '@angular/router';
+import { AlertBannerComponent } from './components/alert-banner/alert-banner.component';
 
 @NgModule({
   declarations: [
@@ -12,18 +18,29 @@ import { FormsModule } from '@angular/forms';
     StylizeDirective,
     OrderByPipe,
     LoaderComponent,
+    GoBackComponent,
+    GoToCartComponent,
+    PathNotFoundComponent,
+    AlertBannerComponent,
+  ],
+  providers: [
+    CartService,
   ],
   exports: [
     HighlightDirective,
     StylizeDirective,
     OrderByPipe,
     LoaderComponent,
+    GoToCartComponent,
+    GoBackComponent,
+    AlertBannerComponent,
     CommonModule,
     FormsModule
   ],
   imports: [
     CommonModule,
-    FormsModule
+    FormsModule,
+    RouterLink
   ]
 })
 export class SharedModule { }